Attempting a variant that imports as a scene rather than a custom JSON format.
